\name{plot.free.energy}
\Rdversion{1.1}
\alias{plot.free.energy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Plot the Free Energy v.s. Time
}
\description{
  Plot the free.energy object generate by the function \code{\link{free.energy}}.
}
\usage{
\method{plot}{free.energy} (x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    an free.energy object generate by the function \code{\link{free.energy}}
  }

  \item{...}{
    additional arguments, passed to \code{\link{plot}}
  }
}

\author{
Hai Qian, Electrical Engineering Department, UCLA
}

\seealso{
  \code{\link{multispin.evol}} for examples
  
  \code{\link{plot.evolution}} is also able to plot free energy of the
  micro-spin system
}


