\name{print.evolution}
\Rdversion{1.1}
\alias{print.evolution}

\title{
  Summarize Some Properties of the Simulation
}
\description{
  A summary of some important properties of the simulation. More is to
  be added. Same as \code{link{summary.evolution}} and \code{\link{print.summary.evolution}}
}
\usage{
\method{print}{evolution} (x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    the simulation result generated by \code{\link{multispin.evol}}
  }

  \item{...}{
    additional arguments, passed to \code{\link{print}}
  }

}

\value{
  The result is a list, whose elements are attributes of \code{g}:
  \code{n}, \code{time}, \code{dt}, \code{call}. The result is defined
  as a "summary.evolution" object.
}

\author{
  Hai Qian, Electrical Engineering Department, UCLA
}

\seealso{
  \code{\link{multispin.evol}}, \code{\link{summary.evolution}}, \code{\link{print.summary.evolution}},
}

