\name{print.summary.evolution}
\Rdversion{1.1}
\alias{print.summary.evolution}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Print the Summary of the Simulation Result
}
\description{
  Print the summary of the simulation result
}
\usage{
\method{print}{summary.evolution} (x, \dots)
}

\arguments{
  \item{x}{
    a summary.evolution object
  }

  \item{...}{
    additional arguments, passed to \code{\link{print}}
  }
}
\author{
  Hai Qian, Electrical Engineering Department, UCLA
}

\seealso{
  \code{\link{summary.evolution}}, \code{\link{print.evolution}}
}
