\name{replot.animation}
\Rdversion{1.1}
\alias{replot.animation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Replot the Animation Generated by component.animation
}
\description{
  The function \code{\link{component.animation}} creats an animation for
  one specific component of all the micro-spins as a new ggobi
  object. This function replots the animation in the same windows.
}
\usage{
replot.animation(ani)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ani}{
    the result of the function \code{\link{component.animation}}
  }
}
\details{
  Sometimes the animation created by \code{\link{component.animation}}
  is not satisfying because of the wrong scales. One can adjust the
  parameters for the animation in the ggobi windows and then replot it
  in the same windows.
}

\references{
  check the reference in \code{\link{component.animation}}
}
\author{
  Hai Qian, Electrical Engineering Department, UCLA
}


\seealso{
  \code{\link{multispin.evol}} to generate simulation result (evolution
  object) and checkout the examples there
  
  \code{\link{component.animation}} to create a new ggobi animation
}

